Inativar disciplinas com solicitações associadas e mais toasts

+++ Dev2/users/views/disciplineView.py
@@ -1,6 +1,8 @@
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.response import Response
 from rest_framework import status
+
+from users.models.forms import KnowledgeCertification, RecognitionOfPriorLearning
 from ..models.discipline import Disciplines
 from ..serializers.disciplineSerializer import DisciplineSerializer
 from rest_framework.permissions import IsAuthenticated
@@ -60,5 +62,21 @@ def discipline_detail(request, pk):
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     elif request.method == 'DELETE':
-        discipline.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        has_recognition_forms = RecognitionOfPriorLearning.objects.filter(discipline=discipline).exists()
+        has_certification_forms = KnowledgeCertification.objects.filter(discipline=discipline).exists()
+
+        if has_recognition_forms or has_certification_forms:
+            # Se houver associações, apenas inativa a disciplina
+            if discipline.is_active:
+                discipline.is_active = False
+                discipline.save()
+                return Response({"detail": "Disciplina foi inativada pois está associada a requisições."},
+                                status=status.HTTP_200_OK)
+            else:
+                discipline.is_active = True
+                discipline.save()
+                return Response({"detail": "Disciplina foi ativa."}, status=status.HTTP_200_OK)
+        else:
+            discipline.delete()
+            # Retorna 204 No Content para indicar sucesso na exclusão
+            return Response({"detail": "Disciplina foi excluida."}, status=status.HTTP_200_OK)
